requirements.txt
flask
flask-cors
openai
langchain
chromadb
faiss-cpu
python-dotenv
pypdf
